style(windows): silence remaining MSVC C4305/C4244 warnings (follow-up)
Continuation of #14. Mechanical pass over the ~161 warning sites that
were skipped in the first batch. Improved the script to:

  - Handle assignment warnings whose column points into the RHS
    expression (scan backward for `=`).
  - Recognize template-dependent source types (`_Ty` from
    `std::vector<double>::operator[]`).

Also added a tiny targeted pass that suffixes `f` to double literals
inside `_PS_CONST(...)` SSE constant-table macros in
`inc/fast-GB/SSEApproxMath.h`.

Files touched (sites fixed):
  - src/f2dock/Docking.cpp                 48
  - inc/fast-GB/SSEApproxMath.h            21 (_PS_CONST literals)
  - src/PG-range/PG.cc                     15
  - src/fast-hydro/pseudoGsol.cpp          15
  - src/fast-PQ/PairingHeap.cpp             9
  - src/GB-rerank/pairingHeap.cpp           9
  - src/f2dock/F2Dock.cpp                   7
  - src/GB-rerank/GB-rerank.cpp             6
  - src/fast-resCont/resContFilter.cpp      4
  - inc/PG-range/PG.h                       1

~31 warning sites remain (complex multi-line expressions, std::printf
format-argument conversions, and a few corner cases). They are deferred
to a future manual pass.

libcvc warnings (`build/_deps/libcvc-src/...`) remain untouched and will
be addressed upstream.

+++ src/GB-rerank/GB-rerank.cpp
@@ -385,9 +385,9 @@ bool rerankF2DockOutput(PARAMS_IN *params, MOLECULE_INFO *staticMol,
       PG *staticPG = new PG(10.0, xlatePG, 5.0);
 
       for (int i = 0; i < staticMol->numQPoints; i++) {
-        staticQPoints[i].x = staticMol->qPoints[7 * i + 0];
-        staticQPoints[i].y = staticMol->qPoints[7 * i + 1];
-        staticQPoints[i].z = staticMol->qPoints[7 * i + 2];
+        staticQPoints[i].x = (float)(staticMol->qPoints[7 * i + 0]);
+        staticQPoints[i].y = (float)(staticMol->qPoints[7 * i + 1]);
+        staticQPoints[i].z = (float)(staticMol->qPoints[7 * i + 2]);
 
         staticPG->addPoint(&staticQPoints[i]);
       }
@@ -487,9 +487,9 @@ bool rerankF2DockOutput(PARAMS_IN *params, MOLECULE_INFO *staticMol,
         PG *movingPG = new PG(10.0, xlatePG, 5.0);
 
         for (int j = 0; j < movingMol->numQPoints; j++) {
-          movingQPoints[j].x = qPoints[7 * j + 0];
-          movingQPoints[j].y = qPoints[7 * j + 1];
-          movingQPoints[j].z = qPoints[7 * j + 2];
+          movingQPoints[j].x = (float)(qPoints[7 * j + 0]);
+          movingQPoints[j].y = (float)(qPoints[7 * j + 1]);
+          movingQPoints[j].z = (float)(qPoints[7 * j + 2]);
 
           movingPG->addPoint(&movingQPoints[j]);
         }
